pyproject.toml
[tool.poetry]
name = "yt-subs-bot"
version = "0.1.0"
description = ""
authors = ["michaelfmnk <michael.fmnk@gmail.com>"]
readme = "README.md"
packages = [{include = "yt_subs_bot"}]

[tool.poetry.scripts]
bot = "yt_subs_bot.app:main"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.3"
requests = "^2.31.0"
pymongo = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
